# home/src/IO.Swagger/Controllers/DeviceApi.cs
/*
 * home
 *
 * The API for the Home Starter project
 *
 * OpenAPI spec version: 1.0.3
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Collections.Generic;
using System.Linq;
using System.Net;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.WebUtilities;
using Microsoft.Extensions.Logging;
using Microsoft.Extensions.Primitives;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    public class DeviceApiController : Controller
    { 
        /// <summary>
        /// 
        /// </summary>
        /// <remarks>returns all registered devices</remarks>
        /// <param name="skip">number of records to skip</param>
        /// <param name="limit">max number of records to return</param>
        /// <response code="200">All the devices</response>
        [HttpGet]
        [Route("/motta/home/1.0.3/devices")]
        [ValidateModelState]
        [SwaggerOperation("GetDevices")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<string>), description: "All the devices")]
        public virtual IActionResult GetDevices([FromQuery]int? skip, [FromQuery]int? limit)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<string>));

            string exampleJson = null;
            exampleJson = "[ \"http://10.0.0.225:8080\", \"http://10.0.0.225:8080\" ]";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<List<string>>(exampleJson)
            : default(List<string>);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        
        /// <param name="device"></param>
        /// <response code="200">successfully registered device</response>
        [HttpPost]
        [Route("/motta/home/1.0.3/devices")]
        [ValidateModelState]
        [SwaggerOperation("Register")]
        public virtual IActionResult Register([FromBody]DeviceRegistrationInfo device)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);


            throw new NotImplementedException();
        }
    }
}